further clear up and refactor
