+++ helix-db/src/helixc/analyzer/methods/infer_expr_type.rs
@@ -1276,9 +1276,5 @@ pub(crate) fn infer_expr_type<'a>(
                 })),
             )
         }
-        _ => {
-            println!("Unknown expression: {expr:?}");
-            todo!()
-        }
     }
 }
