+++ helix-db/src/helixc/parser/types.rs
@@ -476,7 +476,6 @@ pub enum ExpressionType {
     BooleanLiteral(bool),
     ArrayLiteral(Vec<Expression>),
     Exists(ExistsExpression),
-    BatchAddVector(BatchAddVector),
     AddVector(AddVector),
     AddNode(AddNode),
     AddEdge(AddEdge),
@@ -507,7 +506,6 @@ impl Debug for ExpressionType {
             ExpressionType::BooleanLiteral(b) => write!(f, "{b}"),
             ExpressionType::ArrayLiteral(a) => write!(f, "Array({a:?})"),
             ExpressionType::Exists(e) => write!(f, "Exists({e:?})"),
-            ExpressionType::BatchAddVector(bav) => write!(f, "BatchAddVector({bav:?})"),
             ExpressionType::AddVector(av) => write!(f, "AddVector({av:?})"),
             ExpressionType::AddNode(an) => write!(f, "AddNode({an:?})"),
             ExpressionType::AddEdge(ae) => write!(f, "AddEdge({ae:?})"),
@@ -531,7 +529,6 @@ impl Display for ExpressionType {
             ExpressionType::BooleanLiteral(b) => write!(f, "{b}"),
             ExpressionType::ArrayLiteral(a) => write!(f, "Array({a:?})"),
             ExpressionType::Exists(e) => write!(f, "Exists({e:?})"),
-            ExpressionType::BatchAddVector(bav) => write!(f, "BatchAddVector({bav:?})"),
             ExpressionType::AddVector(av) => write!(f, "AddVector({av:?})"),
             ExpressionType::AddNode(an) => write!(f, "AddNode({an:?})"),
             ExpressionType::AddEdge(ae) => write!(f, "AddEdge({ae:?})"),
